requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp
aiosignal
async-timeout
azure-core
azure-cosmos 
azure-functions
azure-storage-blob
beautifulsoup4
isodate
multidict
numpy
pandas
pyodbc
python-dotenv
requests
SQLAlchemy
urllib3
requests